Update README.md
Blood Bank Management System is a web-based platform designed to manage blood donations, requests, and blood banks in a smart and efficient way. The system allows donors to register and request donations, patients to submit blood requests, and administrators to manage blood banks, donations, and users. The system supports real-time notifications, and provides comprehensive dashboards for both donors and administrators. It enhances accessibility to blood and helps save lives by connecting the right donors to the right patients at the right time.